requirements.txt
# Core numerical + plotting
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12

# Dimensionality reduction (UMAP / t-SNE figures)
umap-learn>=0.5

# Deep learning backend (used by a subset of embedding paths)
torch>=2.0

# Diversity metrics
vendi-score>=0.0.3
sacrebleu>=2.3                # Self-BLEU in compute_vendi_extended
nltk>=3.8                      # tokenization / n-gram helpers

# LLM client (for proposal generation + OpenAI embeddings)
openai>=1.0

# Config + misc
PyYAML>=6.0
tqdm>=4.65

# External: the simulation/ runners additionally require
#   - AgentVerse: https://github.com/OpenBMB/AgentVerse
#   - AI-Scientist: https://github.com/SakanaAI/AI-Scientist
# See README for setup details.
